Sort "Ensure crypto" block
